+++ .github/workflows/benchmarks.yml
@@ -49,7 +49,9 @@ jobs:
 
       # Run benchmarks
       - name: Run cargo bench
-        run: cargo bench --features test-utils,experimental-db
+        run: |
+          cargo bench --package floresta-chain --no-default-features --features test-utils,kv-chainstore
+          cargo bench --package floresta-chain --no-default-features --features test-utils,flat-chainstore
 
       # Save cache only if the previous steps succeeded and there was not an exact cache key match
       # This happens everytime we modify any `cargo.lock` or `cargo.toml`, or each two weeks (caching recent changes)
