+++ crates/floresta-chain/Cargo.toml
@@ -21,7 +21,7 @@ crate-type = ["cdylib", "rlib"]
 rustreexo = "0.4"
 sha2 = "^0.10.6"
 log = "0.4"
-kv = "0.24.0"
+kv = { version = "0.24.0", optional = true }
 bitcoin = { version = "0.32", features = [
     "serde",
 ], default-features = false }
@@ -47,9 +47,10 @@ default = []
 bitcoinconsensus = ["bitcoin/bitcoinconsensus", "dep:bitcoinconsensus"]
 metrics = ["dep:metrics"]
 test-utils = ["dep:serde"]
-experimental-db = ["dep:memmap2", "dep:lru"]
+flat-chainstore = ["dep:memmap2", "dep:lru"]
+kv-chainstore = ["dep:kv"]
 
 [[bench]]
 name = "chain_state_bench"
 harness = false
-required-features = ["test-utils", "experimental-db"]
+required-features = ["test-utils"]
