+++ crates/floresta-chain/src/lib.rs
@@ -19,8 +19,10 @@ pub use pruned_utreexo::chain_state::*;
 pub use pruned_utreexo::chainparams::*;
 pub use pruned_utreexo::chainstore::*;
 pub use pruned_utreexo::error::*;
-#[cfg(feature = "experimental-db")]
+#[cfg(feature = "flat-chainstore")]
 pub use pruned_utreexo::flat_chain_store::*;
+#[cfg(feature = "kv-chainstore")]
+pub use pruned_utreexo::kv_chainstore::*;
 pub use pruned_utreexo::udata::*;
 pub use pruned_utreexo::ChainBackend;
 pub use pruned_utreexo::Notification;
