+++ crates/floresta-chain/src/pruned_utreexo/error.rs
@@ -189,5 +189,6 @@ impl Display for BlockchainError {
     }
 }
 
+#[cfg(feature = "kv-chainstore")]
 impl DatabaseError for kv::Error {}
 impl Error for BlockchainError {}
