+++ crates/floresta-chain/src/pruned_utreexo/mod.rs
@@ -13,10 +13,12 @@ pub mod chain_state;
 pub mod chain_state_builder;
 pub mod chainparams;
 pub mod chainstore;
+#[cfg(feature = "kv-chainstore")]
+pub mod kv_chainstore;
 #[macro_use]
 pub mod error;
 pub mod consensus;
-#[cfg(feature = "experimental-db")]
+#[cfg(feature = "flat-chainstore")]
 pub mod flat_chain_store;
 pub mod partial_chain;
 pub mod udata;
