+++ crates/floresta-chain/src/pruned_utreexo/udata.rs
@@ -642,7 +642,7 @@ pub mod proof_util {
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "kv-chainstore"))]
 mod test {
     extern crate std;
