+++ crates/floresta-electrum/Cargo.toml
@@ -35,3 +35,4 @@ thiserror = "1.0"
 rand = "0.8.5"
 rcgen = "0.13"
 zstd = "0.13.3"
+floresta-chain = { path = "../floresta-chain", features = ["kv-chainstore"] }
