+++ crates/floresta-wire/Cargo.toml
@@ -39,6 +39,7 @@ rustls = "0.23.27"
 [dev-dependencies]
 zstd = "0.13.3"
 hex = "0.4.3"
+floresta-chain = { path = "../floresta-chain", features = ["flat-chainstore"] }
 
 [features]
 default = []
